Agents/requirements.txt
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
google-genai>=0.2.0
groq>=0.4.0
ortools>=9.7.0
python-dotenv>=1.0.0
pandas>=2.1.0
jinja2>=3.1.2
